+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 -----------------
 
 - [PP-697] Fixed: unrounded zero-like spent balance amount are now rounded.
+- [PP-697] Fixed: meal order amount computation.
+- [PP-697] Added: pseudo-unit test on meal order computation through a test endpoint.
 
 3.1.3
 -----------------